pyproject.toml
[tool.poetry]
name = "streamlit-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "streamlit_project"}]

[tool.poetry.dependencies]
python = ">=3.8.5,<3.9.7"
streamlit = "^1.24.1"
pygame = "^2.5.0"
mido = "^1.2.10"
torch = {url="https://download.pytorch.org/whl/cu110/torch-1.7.0%2Bcu110-cp38-cp38-linux_x86_64.whl"}
torchvision = {url="https://download.pytorch.org/whl/cu110/torchvision-0.8.1%2Bcu110-cp38-cp38-linux_x86_64.whl"}
pafy = "^0.5.5"
validators = "^0.20.0"
youtube-dl = "^2021.12.17"
ultralytics = "^8.0.132"
librosa = "^0.10.0.post2"
pretty-midi = "^0.2.10"
pytube = "^15.0.0"
ffmpeg-python = "^0.2.0"
pyfluidsynth = "^1.3.2"
opencv-python = "^4.8.0.74"
music21 = "8.3.0"
streamlit-ext = "^0.1.7"
streamlit-player = "^0.1.5"
streamlit-option-menu = "^0.3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
